Refactor venue structure closes #231

+++ internal/fixture/processor_test.go
@@ -47,7 +47,7 @@ func TestProcess(t *testing.T) {
 	t.Run("inserts new fixture", func(t *testing.T) {
 		done := make(chan bool)
 
-		seasonRepo.On("Ids").Return([]int{123}, nil)
+		seasonRepo.On("Ids").Return([]int64{123}, nil)
 		fixtureRepo.On("ById", uint64(34)).Return(&model.Fixture{}, errors.New("not found"))
 		fixtureRepo.On("Insert", mock.Anything).Return(nil)
 		fixtureRepo.AssertNotCalled(t, "Update", mock.Anything)
@@ -68,7 +68,7 @@ func TestProcess(t *testing.T) {
 	t.Run("inserts new fixture", func(t *testing.T) {
 		done := make(chan bool)
 
-		seasonRepo.On("CurrentSeasonIds").Return([]int{123}, nil)
+		seasonRepo.On("CurrentSeasonIds").Return([]int64{123}, nil)
 		fixtureRepo.On("ById", uint64(34)).Return(&model.Fixture{}, errors.New("not found"))
 		fixtureRepo.On("Insert", mock.Anything).Return(nil)
 		fixtureRepo.AssertNotCalled(t, "Update", mock.Anything)
@@ -79,7 +79,7 @@ func TestProcess(t *testing.T) {
 		done := make(chan bool)
 
 		f := newFixture(34)
-		seasonRepo.On("CurrentSeasonIds").Return([]int{123}, nil)
+		seasonRepo.On("CurrentSeasonIds").Return([]int64{123}, nil)
 		fixtureRepo.On("ById", 34).Return(f, nil)
 		fixtureRepo.On("Update", &f).Return(nil)
 		fixtureRepo.AssertNotCalled(t, "Insert", mock.Anything)
@@ -113,20 +113,20 @@ func (m mockSeasonRepository) Update(c *model.Season) error {
 	return args.Error(0)
 }
 
-func (m mockSeasonRepository) Id(id int) (*model.Season, error) {
+func (m mockSeasonRepository) Id(id int64) (*model.Season, error) {
 	args := m.Called(id)
 	c := args.Get(0).(*model.Season)
 	return c, args.Error(1)
 }
 
-func (m mockSeasonRepository) Ids() ([]int, error) {
+func (m mockSeasonRepository) Ids() ([]int64, error) {
 	args := m.Called()
-	return args.Get(0).([]int), args.Error(1)
+	return args.Get(0).([]int64), args.Error(1)
 }
 
-func (m mockSeasonRepository) CurrentSeasonIds() ([]int, error) {
+func (m mockSeasonRepository) CurrentSeasonIds() ([]int64, error) {
 	args := m.Called()
-	return args.Get(0).([]int), args.Error(1)
+	return args.Get(0).([]int64), args.Error(1)
 }
 
 type mockFixtureRepository struct {
